+++ src/device/tsp/Tsp/Script/ExportExtensions/Methods_export.cs
@@ -2,6 +2,26 @@ namespace cc.isr.VI.Tsp.Script.ExportExtensions;
 
 public static partial class ExportExtensionsMethods
 {
+    /// <summary>   A <see cref="string"/> extension method ensure that the script source ends with only one <see cref="Environment.NewLine"/> ending. </summary>
+    /// <remarks>   2025-05-03. <para>
+    /// This ensures that source code exporting anc compression can be validated as both end with a single line terminations.. </para>
+    /// </remarks>
+    /// <exception cref="ArgumentNullException">    Thrown when one or more required arguments are
+    ///                                             null. </exception>
+    /// <param name="scriptSource"> The script source. </param>
+    /// <returns>   A string. </returns>
+    public static string TrimMultipleLineEndings( this string scriptSource )
+    {
+        if ( string.IsNullOrWhiteSpace( scriptSource ) ) throw new ArgumentNullException( nameof( scriptSource ) );
+
+        // ensure that the source ends with a single new line character.
+        while ( scriptSource.EndsWith( $"{Environment.NewLine}{Environment.NewLine}" ) )
+        {
+            scriptSource = scriptSource.TrimEnd( '\r', '\n' );
+        }
+        return $"{scriptSource.TrimEnd( '\r', '\n' )}{Environment.NewLine}";
+    }
+
     /// <summary>
     /// A <see cref="TextReader"/> extension method that exports a script to file in Windows end-of-
     /// line format.
@@ -13,7 +33,9 @@ public static partial class ExportExtensionsMethods
     /// from the source file became the last line of the destination file.<c>
     /// "\91\201\2\185\92I\2\186\92\201\2\187\93I\2\188\93\201\2\189^", "\91\201\2\185\92I\2\186\92\20
     /// </c>
-    /// </para>
+    /// </para><para>
+    /// Note that the output file ends with new line characters. </para>
+    /// 
     /// </remarks>
     /// <exception cref="ArgumentNullException">        Thrown when one or more required arguments
     ///                                                 are null. </exception>
@@ -58,6 +80,8 @@ public static partial class ExportExtensionsMethods
     /// from the source file became the last line of the destination file.<c>
     /// "\91\201\2\185\92I\2\186\92\201\2\187\93I\2\188\93\201\2\189^", "\91\201\2\185\92I\2\186\92\20
     /// </c>
+    /// </para> <para>
+    /// Note that the output ends with a line characters.
     /// </para>
     /// </remarks>
     /// <exception cref="ArgumentNullException">        Thrown when one or more required arguments
@@ -104,6 +128,9 @@ public static partial class ExportExtensionsMethods
 
         if ( !overWrite && File.Exists( filePath ) )
             throw new InvalidOperationException( $"The script source cannot be exported because the file '{filePath}' exists." );
+
+        // ensure that the source ends with a single line termination.
+        scriptSource = scriptSource.TrimMultipleLineEndings();
 
         // Note that the actual output is not validated against the input because the line endings might change.
         // Rather, what is validated is that the reader and writer are able to process the script source.